+++ crates/rpc/src/server/api.rs
@@ -17,13 +17,14 @@ use miden_node_utils::limiter::{
     QueryParamNullifierLimit,
     QueryParamStorageMapKeyTotalLimit,
 };
-use miden_protocol::batch::ProvenBatch;
+use miden_protocol::batch::ProposedBatch;
 use miden_protocol::block::{BlockHeader, BlockNumber};
 use miden_protocol::note::{Note, NoteRecipient, NoteScript};
 use miden_protocol::transaction::{OutputNote, ProvenTransaction, ProvenTransactionBuilder};
 use miden_protocol::utils::serde::{Deserializable, Serializable};
 use miden_protocol::{MIN_PROOF_SECURITY_LEVEL, Word};
 use miden_tx::TransactionVerifier;
+use miden_tx_batch_prover::LocalBatchProver;
 use tonic::{IntoRequest, Request, Response, Status};
 use tracing::{debug, info};
 use url::Url;
@@ -387,47 +388,81 @@ impl api_server::Api for RpcService {
 
     /// Deserializes the batch, strips MAST decorators from full output note scripts, rebuilds
     /// the batch, then forwards it to the block producer.
-    async fn submit_proven_batch(
+    async fn submit_batch(
         &self,
-        request: tonic::Request<proto::transaction::ProvenTransactionBatch>,
+        request: tonic::Request<proto::transaction::TransactionBatch>,
     ) -> Result<tonic::Response<proto::blockchain::BlockNumber>, Status> {
         let Some(block_producer) = &self.block_producer else {
             return Err(Status::unavailable("Batch submission not available in read-only mode"));
         };
 
-        let mut request = request.into_inner();
+        let request = request.into_inner();
 
-        let batch = ProvenBatch::read_from_bytes(&request.encoded)
-            .map_err(|err| Status::invalid_argument(err.as_report_context("invalid batch")))?;
+        let batch = ProposedBatch::read_from_bytes(&request.proposed_batch).map_err(|err| {
+            Status::invalid_argument(err.as_report_context("invalid proposed_batch"))
+        })?;
 
-        // Build a new batch with output notes' decorators removed
-        let stripped_outputs: Vec<OutputNote> =
-            strip_output_note_decorators(batch.output_notes().iter()).collect();
+        // Perform this check here since its cheap. If this passes we can safely zip inputs and
+        // transactions.
+        if request.transaction_inputs.len() != batch.transactions().len() {
+            return Err(Status::invalid_argument(format!(
+                "Number of inputs {} does not match number of transaction {} in batch",
+                request.transaction_inputs.len(),
+                batch.transactions().len()
+            )));
+        }
 
-        let rebuilt_batch = ProvenBatch::new(
-            batch.id(),
-            batch.reference_block_commitment(),
-            batch.reference_block_num(),
-            batch.account_updates().clone(),
-            batch.input_notes().clone(),
-            stripped_outputs,
-            batch.batch_expiration_block_num(),
-            batch.transactions().clone(),
-        )
-        .map_err(|e| Status::invalid_argument(e.to_string()))?;
-
-        request.encoded = rebuilt_batch.to_bytes();
-
-        // Only allow deployment transactions for new network accounts
-        for tx in batch.transactions().as_slice() {
-            if tx.account_id().is_network() && !tx.initial_state_commitment().is_empty() {
+        // Only allow deployment transactions for new network accounts.
+        for tx in batch.transactions() {
+            if tx.account_id().is_network()
+                && !tx.account_update().initial_state_commitment().is_empty()
+            {
                 return Err(Status::invalid_argument(
                     "Network transactions may not be submitted by users yet",
                 ));
             }
         }
 
-        block_producer.clone().submit_proven_batch(request).await
+        // Verify batch transaction proofs.
+        let proof = LocalBatchProver::new(MIN_PROOF_SECURITY_LEVEL).prove(batch.clone()).map_err(
+            |err| Status::invalid_argument(err.as_report_context("proposed block proof failed")),
+        )?;
+        // Verify the reference header matches the canonical chain.
+        let reference_header = self
+            .get_block_header_by_number(Request::new(proto::rpc::BlockHeaderByNumberRequest {
+                block_num: proof.reference_block_num().as_u32().into(),
+                include_mmr_proof: false.into(),
+            }))
+            .await?
+            .into_inner()
+            .block_header
+            .expect("store should always send block header");
+        let reference_commitment: Word = reference_header
+            .chain_commitment
+            .expect("store should always fill block header")
+            .try_into()
+            .expect("store Word should be okay");
+        if reference_commitment != proof.reference_block_commitment() {
+            return Err(Status::invalid_argument(format!(
+                "batch reference commitment {} at block {} does not match canonical chain's commitemnt of {}",
+                proof.reference_block_num(),
+                proof.reference_block_commitment(),
+                reference_commitment
+            )));
+        }
+
+        // Submit each transaction to the validator.
+        //
+        // SAFETY: We checked earlier that the two iterators are the same length.
+        for (tx, inputs) in batch.transactions().iter().zip(&request.transaction_inputs) {
+            let request = proto::transaction::ProvenTransaction {
+                transaction: tx.to_bytes(),
+                transaction_inputs: inputs.clone().into(),
+            };
+            self.validator.clone().submit_proven_transaction(request).await?;
+        }
+
+        block_producer.clone().submit_batch(request).await
     }
 
     // -- Status & utility endpoints ----------------------------------------------------------